+++ hfie/endpoint.py
@@ -1,10 +1,9 @@
-import json
-
 import requests
 import typer
 
+from hfie.api import call
 from hfie.settings import Settings
-from hfie.utils import format_table
+from hfie.utils import format_table, load_json
 
 settings = Settings()
 
@@ -64,10 +63,14 @@ def create(
 ):
     """
     Create an endpoint
+
+    Args:
+        data (str): Path to JSON data to create the endpoint
     """
-    with open(data) as f:
-        data = json.load(f)
-    r = requests.post(f"{settings.endpoint_url}", headers=headers, json=data)
+
+    data = load_json(data)
+    r = call(settings.endpoint_url, headers, data, call="post")
+    typer.secho("Endpoint created successfully", fg=typer.colors.GREEN)
     typer.echo(r.json())
 
 
@@ -79,9 +82,10 @@ def update(
     """
     Update an endpoint
     """
-    with open(data) as f:
-        data = json.load(f)
-    r = requests.put(f"{settings.endpoint_url}/{name}", headers=headers, json=data)
+    data = load_json(data)
+
+    r = call(f"{settings.endpoint_url}/{name}", headers, data, call="put")
+    typer.secho("Endpoint updated successfully", fg=typer.colors.GREEN)
     typer.echo(r.json())
 
 
@@ -95,24 +99,20 @@ def delete(
     """
     Delete an endpoint
     """
+
     if not force:
         delete_endpoint = typer.confirm(
-            f"Are you sure you want to delete endpoint. Use --force to override"
+            "Are you sure you want to delete endpoint. Use --force to override"
         )
+
         if not delete_endpoint:
             typer.echo("Not deleting endpoint")
             raise typer.Abort()
 
     if force or delete_endpoint:
-        r = requests.delete(f"{settings.endpoint_url}/{name}", headers=headers)
-        typer.echo(r.json())
 
-
-def get_info(name: str):
-
-    r = requests.get(f"{settings.endpoint_url}/{name}", headers=headers)
-
-    return r.json()
+        call(f"{settings.endpoint_url}/{name}", headers=headers, data={}, call="delete")
+        typer.secho("Endpoint deleted successfully", fg=typer.colors.GREEN)
 
 
 @app.command()
@@ -123,7 +123,9 @@ def info(
     """
     Get info about an endpoint
     """
-    info = get_info(name)
+
+    r = call(f"{settings.endpoint_url}/{name}", headers=headers, data={}, call="get")
+    info = r.json()
 
     if info.get("name"):
 
@@ -162,21 +164,43 @@ def logs(name: str = typer.Argument(..., help="Endpoint name")):
     """
     Get logs about an endpoint
     """
-    r = requests.get(f"{settings.endpoint_url}/{name}/logs", headers=headers)
+    r = call(
+        f"{settings.endpoint_url}/{name}/logs", headers=headers, data={}, call="get"
+    )
     typer.echo(r.content)
 
 
 @app.command()
 def test(
     name: str = typer.Argument(..., help="Endpoint name"),
-    inputs: str = typer.Argument(..., help="Input to send the model."),
+    inputs: str = typer.Argument(None, help="Input to send the model."),
+    input_file: str = typer.Option(
+        None, help="Path to JSON file containing queries to send to the model."
+    ),
 ):
     """
     Test an endpoint
     """
 
-    info = get_info(name)
+    if not inputs and not input_file:
+        typer.secho(
+            "You must provide either an input string or an input JSON containing your queries",
+            fg=typer.colors.RED,
+        )
+        raise typer.Abort()
+
+    # Get the endpoint url from endpoint info
+
+    r = call(f"{settings.endpoint_url}/{name}", headers=headers, data={}, call="get")
+    info = r.json()
     url = info["status"]["url"]
-    data = {"inputs": inputs, "parameters": {"top_k": 10}}
-    r = requests.post(url, headers=headers, json=data)
+
+    if input_file:
+        data = load_json(input_file)
+    else:
+        data = {"inputs": inputs, "parameters": {"top_k": 10}}
+
+    # Send a call to the endpoint
+
+    r = call(url, headers=headers, data=data, call="post")
     typer.echo(r.json())