+++ hfie/utils.py
@@ -1,3 +1,6 @@
+import json
+
+import typer
 import wasabi
 
 
@@ -27,3 +30,21 @@ def format_table(headers: list, *args):
     table = wasabi.table(data, header=headers, divider=True, widths=widths)
 
     return table
+
+
+def load_json(path: str):
+    """
+    Loads a JSON file
+    """
+
+    try:
+        with open(path) as f:
+            data = json.load(f)
+
+        return data
+    except FileNotFoundError:
+        typer.secho(f"File {path} not found", fg=typer.colors.RED)
+        raise typer.Exit(code=1)
+    except json.decoder.JSONDecodeError:
+        typer.secho(f"File {path} is not a valid JSON", fg=typer.colors.RED)
+        raise typer.Exit(code=1)